Added Swagger API with UI

+++ src/main/java/com/michaldurkalec/fw/fastnfurious/domain/Movie.java
@@ -31,7 +31,7 @@ public class Movie {
     public Float getAvgScore() {
         Double avgScore = ratings.stream()
                 .filter(Objects::nonNull)
-                .mapToDouble(Rating::getScore)
+                .mapToInt(Rating::getScore)
                 .average()
                 .orElse(Double.NaN);
         return avgScore.isNaN()
